requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
plotly==5.23.0
pandas==2.2.2
numpy==1.26.4
schwabdev==2.5.1
python-dotenv==1.0.1
dash-daq==0.5.0
websocket-client==1.7.0
scipy==1.13.1